Restructered Equity and added logger functions
