+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
-ignore = E401,E402,E501,W503,E712,F401, F405, F403, E203
+ignore = E401,E402,E501,W503,E712,F401, F405, F403, E203, E231, E241, E202
 max-line-length = 120
