+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
     rev: "22.3.0"  # Updated to the correct latest version
     hooks:
       - id: black
-        language_version: python3.11  # Ensure it uses Python 3.11
+        language_version: python3.12  # Ensure it uses Python 3.11
 
 
   - repo: https://github.com/pycqa/flake8
